Virtual_Tutor_Talk/requirements.txt
streamlit             # Streamlit library
azure-cognitiveservices-speech  # Azure Cognitive Services Speech SDK
openai                # OpenAI Python client library
python-dotenv         # Python dotenv library